requirements.txt
aiohttp==3.7.4.post0
aiosqlite==0.17.0
astunparse==1.6.3
async-timeout==3.0.1
asyncio-dgram==2.1.0
asyncpg==0.25.0
attrs==21.2.0
braceexpand==0.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
dnspython==2.1.0
fonttools==4.28.2
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
idna==3.3
import-expression==1.1.4
instant-api-client==0.1.2
jishaku @ git+https://github.com/ooliver1/jishaku-nextcord@f404fa11c956e8ac37a4e843e26aa0c868531316
kiwisolver==1.3.2
matplotlib==3.5.0
mcstatus==6.5.0
multidict==5.2.0
nextcord==2.0.0a3
nextcord-ext-menus==1.3.2
numpy==1.21.4
packaging==21.3
pafy==0.5.5
Pillow==8.4.0
praw==7.5.0
prawcore==2.3.0
prsaw==0.4.0
psutil==5.8.0
pycparser==2.21
pyfiglet==0.8.post1
PyNaCl==1.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
requests==2.26.0
rfc3986==1.5.0
setuptools-scm==6.3.2
simpcalci==0.0.2
six==1.16.0
sniffio==1.2.0
tabulate==0.8.9
tomli==1.2.2
typing-extensions==3.10.0.2
update-checker==0.18.0
urllib3==1.26.7
wcwidth==0.2.5
websocket-client==1.2.1
yarl==1.7.2
youtube-dl==2021.6.6
